simulators/requirements.txt
paho-mqtt==1.6.1
pymongo==4.6.0
opcua==0.98.13
requests==2.31.0

